requirements.txt
pygame~=2.6.1
pillow~=12.1.0